technology_books.txt
This is a living list of books that I have read and recommend in the area of technology and cybersecurity. This will be updated
over time as I go through my library. The books that do make this list are those that make good resources for the professional or those just getting started.

## Recommended Technology Books: